+++ app/Http/Controllers/ExperienceController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreExperienceRequest;
-use App\Http\Requests\UpdateExperienceRequest;
+use App\Http\Requests\ExperienceRequest;
 use App\Http\Resources\ExperienceResource;
+use App\Models\Education;
 use App\Models\Experience;
 use Illuminate\Support\Facades\Auth;
 
@@ -12,28 +12,29 @@ class ExperienceController extends Controller
 {
     public function index()
     {
-        $experience = Auth::User()->experiences();
-        return response()->json(ExperienceResource::collection($experience));
+        if (Auth::User()->experiences()->isNotEmpty()) {
+            return response()->json(Auth::User()->experiences());
+        }
+        return response()->json(['message' => 'No experiences for You'], 200);
     }
 
-    public function store(StoreExperienceRequest $request)
+    public function store(ExperienceRequest $request)
     {
-        $educate = Experience::create([
-            ...$request->validated(),
+        $experience = Experience::updateOrCreate(['user_id' => auth()->user()->id], [
+            ...$request->all(),
             'user_id' => auth()->user()->id,
         ]);
-        return response()->json(new ExperienceResource($educate));
-    }
-
-    public function update(UpdateExperienceRequest $request, Experience $experience)
-    {
-        $experience->update($request->validated());
         return response()->json(new ExperienceResource($experience));
     }
 
-    public function destroy(Experience $experience)
+
+    public function destroy($id)
     {
-        $experience->delete();
-        return response()->json(['message' => 'Delete Successfully'], 200);
+        $experience = Experience::findOrFail($id);
+        if (auth()->user()->id === $experience->user_id) {
+            $experience->delete();
+            return response()->json(['message' => 'Delete Successfully'], 200);
+        }
+        return response()->json(['error' => 'invalid'], 500);
     }
 }
